refactor: 重构表格跨页行列操作实现
